Commit old script for release via Github
I originally wrote this script back in 2008. It worked back then for
Samba3 with tdbsam backend, but I have no idea if it works now. I am
releasing on Github because of a random E-Mail request for it (and
discovering that our original "devzone" link is dead, which will be
fixed and redirected to the script's new home on Github shortly)